feat: Change layout of queue to select cocktail or open dialog when clicking on the name
Queue time and note about multiple entries for the same cocktail have been moved to an extra line.

When on the search page, the cocktail is shown in the selected cocktail widget when clicking on the queue entry.
If not on the search page, the modal is opened (as when clicking on the eye)